refactored user reducers into a singular singleUser reducer and multiple users reducer
