+++ client/components/Signup/signup.js
@@ -1,7 +1,7 @@
 import React, {Component} from 'react'
 import {connect} from 'react-redux'
 import {NavLink} from 'react-router-dom'
-import {addUserToServer} from '../../reducer/user/user'
+import {addUserToServer} from '../../reducer/singleUser'
 //import './signup.css'
 
 class Signup extends Component {
