+++ client/reducer/singleUser.js
@@ -1,26 +1,23 @@
 import axios from 'axios'
 import history from '../history'
 
-/**
- * ACTION TYPES
- */
 const GET_USER = 'GET_USER'
 const REMOVE_USER = 'REMOVE_USER'
+const SIGNUP_USER = 'SIGNUP_USER'
 
-/**
- * INITIAL STATE
- */
 const defaultUser = {}
 
-/**
- * ACTION CREATORS
- */
 const getUser = user => ({type: GET_USER, user})
+
 const removeUser = () => ({type: REMOVE_USER})
 
-/**
- * THUNK CREATORS
- */
+const signupUser = user => {
+  return {
+    type: SIGNUP_USER,
+    user
+  }
+}
+
 export const me = () => async dispatch => {
   try {
     const res = await axios.get('/auth/me')
@@ -55,12 +52,27 @@ export const logout = () => async dispatch => {
   }
 }
 
-/**
- * REDUCER
- */
+export const addUserToServer = user => {
+  return async dispatch => {
+    try {
+      const {data} = await axios.post('/api/users', {
+        firstName: user.firstName,
+        lastName: user.lastName,
+        email: user.email,
+        password: user.password
+      })
+      dispatch(signupUser(data))
+    } catch (error) {
+      console.error(error)
+    }
+  }
+}
+
 export default function getUserReducer(state = defaultUser, action) {
   switch (action.type) {
     case GET_USER:
+      return action.user
+    case SIGNUP_USER:
       return action.user
     case REMOVE_USER:
       return defaultUser
